refactoring, create visit screen folder
